Remvoe usage of useListData
